Created side class structure with one stub implementation

+++ src/puzzlesolver/Point.java
@@ -1,19 +1,19 @@
 package puzzlesolver;
 
 /**
- * A class to represent immutable points, with integer x and y values.
+ * A class to represent immutable points, with x and y values.
  */
 public class Point {
 
-  public final int x;
-  public final int y;
+  public final double x;
+  public final double y;
 
   /**
    * Constructs a new {@code Point} with the given x and y values.
    * @param x the x value of the point
    * @param y the y value of the point
    */
-  public Point(int x, int y){
+  public Point(double x, double y){
     this.x = x;
     this.y = y;
   }
